pyproject.toml
[tool.poetry]
name = "fjsbot"
version = "0.1.0"
description = ""
authors = ["Andreas Løfsgaard <andreas@lofsgaard.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord = "^2.1.0"
python-dotenv = "^0.21.0"
requests = "^2.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
